refactor API key expiry handling
