+++ packages/ui/src/components/apikeyform/ApiKeyForm.module.css
@@ -93,7 +93,7 @@
   color: rgb(17 24 39);
   font-size: var(--large-text);
   line-height: 28px;
-  font-weight: 600;
+  font-weight: 500;
   margin-top: 12px;
 }
