+++ packages/ui/src/components/information/Information.module.css
@@ -1,5 +1,5 @@
 .information-content {
-  padding: 8px;
+  padding: 4px;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -16,9 +16,9 @@
 .information-message {
   font-size: var(--medium-text);
   margin-bottom: 24px;
-  text-align: center;
   color: var(--description);
   line-height: 1.5;
+  text-align: center;
 }
 
 .button-wrapper {
